refactor command handlers
