+++ internal/langserver/handlers/execute_command.go
@@ -2,46 +2,17 @@ package handlers
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
-	"strconv"
-	"strings"
 
 	"github.com/creachadair/jrpc2/code"
 	lsctx "github.com/hashicorp/terraform-ls/internal/context"
+	"github.com/hashicorp/terraform-ls/internal/langserver/cmd"
+	"github.com/hashicorp/terraform-ls/internal/langserver/handlers/command"
 	lsp "github.com/hashicorp/terraform-ls/internal/protocol"
 )
 
-type executeCommandHandler func(context.Context, commandArgs) (interface{}, error)
-type executeCommandHandlers map[string]executeCommandHandler
-
-const langServerPrefix = "terraform-ls."
-
-var handlers = executeCommandHandlers{
-	prefixCommandName("rootmodules"): executeCommandRootModulesHandler,
-}
-
-func prefixCommandName(name string) string {
-	return langServerPrefix + name
-}
-
-func (h executeCommandHandlers) Names(commandPrefix string) (names []string) {
-	if commandPrefix != "" {
-		commandPrefix += "."
-	}
-	for name := range h {
-		names = append(names, commandPrefix+name)
-	}
-	return names
-}
-
-func (h executeCommandHandlers) Get(name, commandPrefix string) (executeCommandHandler, bool) {
-	if commandPrefix != "" {
-		commandPrefix += "."
-	}
-	name = strings.TrimPrefix(name, commandPrefix)
-	handler, ok := h[name]
-	return handler, ok
+var handlers = cmd.Handlers{
+	cmd.Name("rootmodules"): command.RootModulesHandler,
 }
 
 func (lh *logHandler) WorkspaceExecuteCommand(ctx context.Context, params lsp.ExecuteCommandParams) (interface{}, error) {
@@ -57,83 +28,5 @@ func (lh *logHandler) WorkspaceExecuteCommand(ctx context.Context, params lsp.Ex
 	if !ok {
 		return nil, fmt.Errorf("%w: command handler not found for %q", code.MethodNotFound.Err(), params.Command)
 	}
-	return handler(ctx, parseCommandArgs(params.Arguments))
-}
-
-type commandArgs map[string]interface{}
-
-func (c commandArgs) GetString(variable string) (string, bool) {
-	vRaw, ok := c[variable]
-	if !ok {
-		return "", false
-	}
-	v, ok := vRaw.(string)
-	if !ok {
-		return "", false
-	}
-	return v, true
-}
-
-func (c commandArgs) GetNumber(variable string) (float64, bool) {
-	vRaw, ok := c[variable]
-	if !ok {
-		return 0, false
-	}
-	v, ok := vRaw.(float64)
-	if !ok {
-		return 0, false
-	}
-	return v, true
-}
-
-func (c commandArgs) GetBool(variable string) (bool, bool) {
-	vRaw, ok := c[variable]
-	if !ok {
-		return false, false
-	}
-	v, ok := vRaw.(bool)
-	if !ok {
-		return false, false
-	}
-	return v, true
-}
-
-func parseCommandArgs(arguments []json.RawMessage) commandArgs {
-	args := make(map[string]interface{})
-	if arguments == nil {
-		return args
-	}
-	for _, rawArg := range arguments {
-		var arg string
-		err := json.Unmarshal(rawArg, &arg)
-		if err != nil {
-			// TODO: Log error
-			continue
-		}
-		if arg == "" {
-			continue
-		}
-
-		pair := strings.SplitN(arg, "=", 2)
-		if len(pair) != 2 {
-			continue
-		}
-
-		variable := strings.ToLower(pair[0])
-		value := pair[1]
-		if value == "" {
-			args[variable] = value
-			continue
-		}
-
-		if f, err := strconv.ParseFloat(value, 64); err == nil {
-			args[variable] = f
-		} else if b, err := strconv.ParseBool(value); err == nil {
-			args[variable] = b
-		} else {
-			args[variable] = value
-		}
-
-	}
-	return args
+	return handler(ctx, cmd.ParseCommandArgs(params.Arguments))
 }
